+++ consensus/service/src/api/blockchain_api_service.rs
@@ -185,8 +185,9 @@ mod tests {
     use grpcio::{ChannelBuilder, Environment, Error as GrpcError, Server, ServerBuilder};
     use mc_common::{logger::test_with_logger, time::SystemTimeProvider};
     use mc_consensus_api::consensus_common_grpc::{self, BlockchainApiClient};
-    use mc_transaction_core::{BlockVersion, TokenId};
-    use mc_transaction_core_test_utils::{create_ledger, initialize_ledger, AccountKey};
+    use mc_ledger_db::test_utils::{create_ledger, initialize_ledger};
+    use mc_transaction_core::TokenId;
+    use mc_transaction_core_test_utils::AccountKey;
     use mc_util_grpc::{AnonymousAuthenticator, TokenAuthenticator};
     use rand::{rngs::StdRng, SeedableRng};
     use std::{collections::HashMap, time::Duration};
@@ -226,16 +227,14 @@ mod tests {
             &account_key,
             &mut rng,
         );
+        let last_index = block_entities.last().unwrap().block().index;
 
         let mut expected_response = LastBlockInfoResponse::new();
-        expected_response.set_index(block_entities.last().unwrap().index);
+        expected_response.set_index(last_index);
         expected_response.set_mob_minimum_fee(4000000000);
         expected_response.set_minimum_fees(HashMap::from_iter(vec![(0, 4000000000), (60, 128000)]));
         expected_response.set_network_block_version(*BlockVersion::MAX);
-        assert_eq!(
-            block_entities.last().unwrap().index,
-            ledger_db.num_blocks().unwrap() - 1
-        );
+        assert_eq!(last_index + 1, ledger_db.num_blocks().unwrap());
 
         let mut blockchain_api_service =
             BlockchainApiService::new(ledger_db, authenticator, fee_map, BlockVersion::MAX, logger);
@@ -285,18 +284,16 @@ mod tests {
         let authenticator = Arc::new(AnonymousAuthenticator::default());
         let mut rng: StdRng = SeedableRng::from_seed([1u8; 32]);
         let account_key = AccountKey::random(&mut rng);
-        let block_entities = initialize_ledger(
+        let expected_blocks = initialize_ledger(
             BlockVersion::MAX,
             &mut ledger_db,
             10,
             &account_key,
             &mut rng,
-        );
-
-        let expected_blocks: Vec<blockchain::Block> = block_entities
-            .into_iter()
-            .map(|block_entity| blockchain::Block::from(&block_entity))
-            .collect();
+        )
+        .into_iter()
+        .map(|block_data| blockchain::Block::from(block_data.block()))
+        .collect::<Vec<_>>();
 
         let mut blockchain_api_service = BlockchainApiService::new(
             ledger_db,
@@ -371,18 +368,16 @@ mod tests {
         let authenticator = Arc::new(AnonymousAuthenticator::default());
         let mut rng: StdRng = SeedableRng::from_seed([1u8; 32]);
         let account_key = AccountKey::random(&mut rng);
-        let block_entities = initialize_ledger(
+        let expected_blocks = initialize_ledger(
             BlockVersion::MAX,
             &mut ledger_db,
             10,
             &account_key,
             &mut rng,
-        );
-
-        let expected_blocks: Vec<blockchain::Block> = block_entities
-            .into_iter()
-            .map(|block_entity| blockchain::Block::from(&block_entity))
-            .collect();
+        )
+        .into_iter()
+        .map(|block_data| blockchain::Block::from(block_data.block()))
+        .collect::<Vec<_>>();
 
         let mut blockchain_api_service = BlockchainApiService::new(
             ledger_db,
