+++ consensus/service/src/byzantine_ledger/worker.rs
@@ -924,7 +924,10 @@ mod tests {
         MockScpNode, QuorumSet,
     };
     use mc_crypto_multisig::SignerSet;
-    use mc_ledger_db::MockLedger; // Don't use test_utils::MockLedger.
+    use mc_ledger_db::{
+        test_utils::{create_ledger, create_transaction, initialize_ledger},
+        MockLedger,
+    };
     use mc_ledger_sync::{LedgerSyncError, MockLedgerSync, SCPNetworkState};
     use mc_peers::MockBroadcast;
     use mc_peers_test_utils::MockPeerConnection;
@@ -935,8 +938,7 @@ mod tests {
         TokenId,
     };
     use mc_transaction_core_test_utils::{
-        create_ledger, create_mint_config_tx_and_signers, create_mint_tx_to_recipient,
-        create_transaction, initialize_ledger, mint_config_tx_to_validated,
+        create_mint_config_tx_and_signers, create_mint_tx_to_recipient, mint_config_tx_to_validated,
     };
     use mc_util_metered_channel::{Receiver, Sender};
     use mc_util_metrics::OpMetrics;
