+++ ledger/db/Cargo.toml
@@ -5,36 +5,41 @@ authors = ["MobileCoin"]
 edition = "2021"
 
 [features]
-test_utils = ["rand"]
 # The migration_support feature exposes some internals that should only be used by the `mc-ledger-migration` tool.
 migration_support = []
+test_utils = [
+    "mc-blockchain-test-utils",
+    "mc-transaction-core-test-utils",
+    "mc-transaction-std/test-only"
+]
 
 [dependencies]
 mc-account-keys = { path = "../../account-keys" }
 mc-blockchain-types = { path = "../../blockchain/types" }
+mc-blockchain-test-utils = { path = "../../blockchain/test-utils", optional = true }
 mc-common = { path = "../../common", features = ["log"] }
 mc-crypto-keys = { path = "../../crypto/keys" }
 mc-transaction-core = { path = "../../transaction/core" }
+mc-transaction-core-test-utils = { path = "../../transaction/core/test-utils", optional = true }
+mc-transaction-std = { path = "../../transaction/std" }
 mc-util-from-random = { path = "../../util/from-random" }
 mc-util-lmdb = { path = "../../util/lmdb" }
 mc-util-metrics = { path = "../../util/metrics" }
 mc-util-serial = { path = "../../util/serial", features = ["std"] }
 mc-util-telemetry = { path = "../../util/telemetry" }
+mc-util-test-helper = { path = "../../util/test-helper" }
 
-displaydoc = { version = "0.2", default-features = false }
+displaydoc = "0.2"
 lazy_static = "1.4"
-lmdb-rkv = "0.14.0"
-mockall = "0.11.1"
+lmdb-rkv = "0.14"
+mockall = "0.11"
 prost = { version = "0.10", default-features = false, features = ["prost-derive"] }
-rand = { version = "0.8", optional = true }
-rand_core = "0.6"
+rand = "0.8"
+tempdir = "0.3"
 
 [dev-dependencies]
 mc-blockchain-test-utils = { path = "../../blockchain/test-utils" }
-mc-crypto-keys = { path = "../../crypto/keys" }
 mc-crypto-multisig = { path = "../../crypto/multisig" }
 mc-crypto-rand = { path = "../../crypto/rand" }
 mc-transaction-core-test-utils = { path = "../../transaction/core/test-utils" }
-
-rand = "0.8"
-tempdir = "0.3"
+mc-transaction-std = { path = "../../transaction/std", features = ["test-only"] }
