+++ ledger/db/src/mint_config_store.rs
@@ -435,8 +435,7 @@ pub mod tests {
         mint_config_tx_to_validated as to_validated,
     };
     use mc_util_from_random::FromRandom;
-    use rand::{rngs::StdRng, SeedableRng};
-    use rand_core::RngCore;
+    use rand::{rngs::StdRng, RngCore, SeedableRng};
 
     pub fn init_mint_config_store() -> (MintConfigStore, Environment) {
         let env = get_env();
