+++ transaction/core/test-utils/Cargo.toml
@@ -17,7 +17,6 @@ mc-crypto-multisig = { path = "../../../crypto/multisig" }
 mc-crypto-rand = { path = "../../../crypto/rand" }
 mc-crypto-ring-signature-signer = { path = "../../../crypto/ring-signature/signer" }
 mc-fog-report-validation-test-utils = { path = "../../../fog/report/validation/test-utils" }
-mc-ledger-db = { path = "../../../ledger/db" }
 mc-transaction-core = { path = "../../../transaction/core" }
 mc-transaction-std = { path = "../../../transaction/std", features = ["test-only"] }
 mc-util-from-random = { path = "../../../util/from-random" }
