+++ transaction/core/tests/util/mod.rs
@@ -5,12 +5,15 @@
 #![allow(dead_code)]
 
 use mc_account_keys::PublicAddress;
-use mc_ledger_db::{Ledger, LedgerDB};
-use mc_transaction_core::{tx::Tx, BlockVersion};
-use mc_transaction_core_test_utils::{
-    create_ledger, create_transaction, create_transaction_with_amount_and_comparer,
-    create_transaction_with_amount_and_comparer_and_recipients, initialize_ledger, AccountKey,
+use mc_ledger_db::{
+    test_utils::{
+        create_ledger, create_transaction, create_transaction_with_amount_and_comparer,
+        create_transaction_with_amount_and_comparer_and_recipients, initialize_ledger,
+    },
+    Ledger, LedgerDB,
 };
+use mc_transaction_core::{tx::Tx, BlockVersion};
+use mc_transaction_core_test_utils::AccountKey;
 use mc_transaction_std::{DefaultTxOutputsOrdering, TxOutputsOrdering};
 use mc_util_test_helper::{RngType, SeedableRng};
